+++ MyCapytain/errors.py
@@ -60,3 +60,17 @@ class UndispatchedTextError(Exception, MyCapytainException):
 class UnknownCollection(KeyError, MyCapytainException):
     """ A collection is unknown to its ancestor
     """
+
+class EmptyReference(SyntaxWarning, MyCapytainException):
+    """ Error generated when a duplicate is found in Reference
+    """
+
+    
+class CitationDepthError(UnknownObjectError, MyCapytainException):
+    """ Error generated when the depth of a requested citation is deeper than the citation scheme of the text
+    """
+
+
+class MissingRefsDecl(Exception, MyCapytainException):
+    """ A text has no properly encoded refsDecl
+    """